requirements.txt
flask==2.3.2
google-generativeai==0.3.0
markdown==3.4.4
python-dotenv==1.0.0